api-gateway/requirements.txt

fastapi
uvicorn
httpx
python-jose
